international_debt_analysis_project/requirements.txt
pandas
numpy
matplotlib
streamlit
mysql-connector-python
